+++ CHANGELOG.md
@@ -32,6 +32,7 @@ This release supports Quill v1.0.0+. ⚠️ There are many breaking changes, so 
 - Updated the toolbar alignment section (@clemmy)
 - Updated rendering of toolbar actions (@clemmy)
 - Improved toolbar renderChoices implementation (@zhang-z)
+- Fixed use of `defaultValue` in Toolbar selects
 
 v0.4.1
 ------